Add methods to the daemon to allow device profiling
When profiling scanners, printers and displays, we have to ensure that
any other program requesting profiles from colord for a specific
device is disallowed.

If the calling program calls ProfilingInhibit() on the device that is
about to be profiled, then any subsequent calls to GetProfileForQualifier()
will return no profiles.

The profiling program either has to exit, or call ProfilingUninhibit() to
resume normal device profile matching behaviour.