Add a commit method to the device and profile objects
Sometimes we might want to set the device type and other properties before
it is announced to the world using the DeviceAdded() signal. This allows clients
to filter devices based on certain attributes in the Added signal.

This problem also applies for profiles too, so add this feature for those too.