Change GetProfileForQualifier() into GetProfileForQualifiers() so we can send multiple requests
In the CUPS pstoraster code we need to do 8 calls to GetProfileForQualifier()
to cope with the fallbacks we want to do. If each RTT was to take 25ms, then
this would add 200ms to every document printed, which would be uncool on a
print server.

By making GetProfileForQualifiers() take an array of qualifiers in search
order this can be reduced to one RTT, which isn't even detectable.

This obviously breaks API, but only GCM uses this method at the moment.