Add a 'relation' parameter to Device.AddProfile() to solve some hard to fix bugs
The concept of hard and soft profiles on a device allows to us auto-add profiles based on
EDID or from a service like CUPS, with a lower priority to user-selected profiles.

This ensures that CUPS adding a profile does not overwrite the users default profile,
nor does adding a ICC profile into your home directory change the default just because
the EDID hash matches.

Of course, if a profile is marked as default, then it becomes a hard profile, no matter
on where it came from.

This is obviously an API break, but it's certain to make the logic in colord much
cleaner and more understandable.