Use cmsOpenProfileFromFile() rather than getting the complete profile contents to speed up startup
Using g_file_get_contents() we're doing a lot of IO we don't need -- in the
best case we only need the data in the ICC header rather than all the color
tables.

This patch may actually *increase* the amount of memory used slightly if system
profiles that do not have profile IDs (which can be fixed by cd-fix-profile).

This patch reduces the amount of IO done at startup from 6.10Mb to 2.73Mb on a
system with shared-color-profiles >= 0.1.4.