Allow the client to pass file descriptors out of band to CreateProfile
This functionality allows the session client to pass colord a file descriptor
rather than a filename for profile loading.
This allows the daemon to avoid open()ing the file once again, and also allows
frameworks such as SELinux to be confined to not allow daemons running as root
from even accessing files in the users home directory.

To do this without breaking API the libcolord library (which is running in
process with the application) open's the file specified in the 'Filename'
metadata, and passes this to the daemon.

If an old version of colord is running or the bus daemon is too old and does
not support this fd passing feature, then the session just goes back to the old
way of manually re-opening the file like it always used to.

Signed-off-by: Richard Hughes <richard@hughsie.com>