Add support for reading and writing ICC profile metadata
Using LCMS 2.2 we can read and write the DICT tag. This allows us to write
metadata to the profile filename when repairing and creating, and also allows us
to read that metadata so we can add it to the meta-profile on the DBus interface.

When distros are shipping the yet-to-be-released 2.2 we can remove the GCM EDID
hack that was used for profile matching.