Rework libcolord so the async methods never block
This also changes the behaviour somewhat. Now, a client has to 'connect' to a
sensor, device and profile before the properties can be read.
Relying on doing this automatically in a thread was racy as hell and leads to an
over complicated design that leads some async methods to actualy be sync.

This also changes the API of the DeleteProfile and DeleteDevice to use the
object paths rather than profile and device IDs. This allows the end user to
delete the objects without having to connect to the device and make the API
consistent.

The sync methods remain intact from an ABI point of view, but now call into the
async versions for the implementation.

I think everything is a lot saner this way.