Add a timestamp when making profiles default so that the newest assigned profile is always assigned
This involves keeping track of when each profile was made default and doing the
right thing when profiles come and go. This is important if we add devices and
then change the default profile to something other than the last selected device.

This changes the database schema, but existing mapping databases are
automatically converted. I've also added a nice chunk in the self test program
to try and test all the new code.