Split out the sensors into runtime-loadable shared objects
This allows us to start faster without sensors inserted, as the colord binary
is smaller. This also allows us to install additional sensor types that may not
be shipped in the colord tarball by 3rd party vendors.

NOTE: The dlloaded sensor interface is still in flux and is not yet ABI stable.