Add a new method Client:FindDeviceByProperty to get a device when only a property value is known
This allows us to easily get a device for a specific metadata key value, for
instance "XRANDR_name=lvds1". The alternative is getting all the devices and
searching each one, which is quite inefficient as each device would need to be
connect'ed to to get the list of properties and metadata.

Programs like mutter and compiz-cms can use this new method to easily get the
colord device, which in turn allows them to get the default profile for a
display device.