Define the DATA_source profile metadata item
This allows us to tag the profiles with some data that lets us know where the
data came from. This will be used by gnome-color-manager so that we do not rely
on a hacky heuristic to only warn users about old calibration profiles and not
warn them about old auto-EDID profiles.