Add a configure argument of --enable-fd-fallback
This defaults to true, and allows distros to disable opening of the user
profiles from the daemon.

This in theory is a little safer, and does not upset frameworks like SELinux.

The downside is that testing is made much harder and all clients have to support
FD-passing in DBus messages.