Add functionality to test profile ordering
When a device is added and has profiles added to it, each profile<->device
mapping has a timestamp so that we can order the profiles by the most recent
first.

If we add the profile2 and then profile1 to a device, and then remove the
profiles (like on session exit) we expect profile1 to always be first in the list.
The profiles are currently being added to the device from the database in the
correct order, but if the profiles are added in the wrong order, the last
profile created on the next session login gets promoted into the first place.

This means we sometimes get profile2,profile1 on the session restart rather than
the desired profile1,profile2.

This test currently fails, and the daemon needs to be fixed.