Ensure uid 0 can create devices and profiles even when not on the active console
When colord is running as the root user then this magically works, even though
daemons like cups are not on an active console.

If you switch colord to running as a private user then PolicyKit no longer
implicitly grants the authorisation from the root user, and this means that
printers cannot be registered with colord. By checking for the uid we can grant
an implicit authorisation ourselves before asking PolicyKit.