Add --xorg-gamma to cd-create-profile to create custom gamma ramps
This allows power users to create a custom gamma curve using the same values
and in the same format that was set in xorg.conf.

This of course isn't profiling the screen, but it allows people to use a simple
color profile when they just want to use a fixed gamma scaling factor.