Add profile metadata MAPPING_device_id for automatic profile mapping
This allows us to automatically soft-add a profile for a given device.
This would let us profile a device, and by setting the MAPPING_device_id
metadata key to the device_id we could share the profile with other people with
the same hardware.

If this profile was then installed elsewhere, then it would automatically be
mapped to the device if present, just like it would be done from a PPD file.
This of course can be overriden using the mapping database, which uses hard
mappings.