Fix InstallSystemWide() when running as the colord user
When running as the colord user, we're not able to access files in the users
home directory, and so the file copy fails. Instead; Create a mapped file from
the fd that is passed in when the profile is created, and then dump that mapped
file to a new file in the /var/lib/color system wide location.