Add a Manager.CreateProfileWithFd() method
When we're doing FD passing, the DBus spec says we should pass an 'h' handle
type as the index into the UnixFdList. To avoid breaking API, leave the
CreateProfile() DBus signature as (ssa{ss}) and add a new shim method
of (ssha{ss}). You can actually still pass FDs using either method.

This makes implementations using QtDBus possible, which validates the DBus type
signatures in a way that libdbus or GDBus does not.