Ensure we return NULL for missing device properties
The colord-kde frontend sets a zero length string for missing properties, which
confuses the title logic in gnome-control-center when showing the device.

In the same way we try to detect these 'nullable' strings in the profile code,
add it to the libcolord ChDevice code to make the panel do the right thing.