Add a user suffix to the object path of user-created devices and profiles
This means that two users logged in at the same time can have the same device
and profile IDs, and the uid of the calling process is used to choose the
correct profile.

The session should not prefix the username itself to the device-id as the
device-id is required to be predictable for other clients to query.