Add a 'SensorId' property to the sensor interface
This lets us discover and use a unique ID that we can use when seleting a sensor
to use for calibration. This allows us to support machines with an internal
calibration device *and* and external device like the ColorHug.

Also print the sensor ID when the user does 'colormgr get-sensors'.