Add a session helper that can be used to calibrate the screen
This is an optional D-Bus session activated service and designed to be used by
GNOME and KDE for display calibration.

The D-Bus API should be regarded as experimental and should not be considered
stable like the colord daemon D-Bus API.

An example client program is also included, although is disabled by default as
it requires GNOME libraries like libgnomedesktop and also requires colord-gtk
which would both create a dependancy loop for projects like jhbuild.