Add the ability to 'disable' a device so that color management can be turned off
To achieve this, add a 'Device.Enabled' property and a Device.SetEnabled(bool)
method that controls the 'Device.Profiles' property much like the inhibitor
functionality.

If the enabled state is changed then this is reflected for all users and
persistent across reboots. Devices are enabled by default and changing the
enabled state is a privileged operation.