Recognise the internal sensor on W700 laptops and set a property for the calibration tools
Add a 'Sensor.Embedded' property that is set to True for sensor hardware that is
embedded in the laptop palmrest. Calibration software should fullscreen any
sample widget and emit a sound when the calibration process has completed as the
laptop lid needs to be closed.

In the future, we'll control the required ACPI method to stop the machine
automatically turning off the panel 'in hardware' so that calibration just works.