Check we have shared-color-profiles or a compatible profile set
Using some of the random profiles on distros means we get varied and obscure bugs.

For instance, profiles that do not have a ProfileID pre-computed mean that the
daemon is forced to calculate MD5 sums and is very slow to start which makes
colord look broken when profiling boot speed.

Also, some profile sets do not provide some key profiles that applications
depend on which is hurting my argument that applications should not ship thier
own profiles and instead use a known-good system copy.