Fix floating point number parsing in non C locale environments
atof() is local dependant. Passing 1.0 to atof with in a French environment
leads to segfault in lcms for create-standard-space as the primaries are all
equal to 0 due to the dot not meaning decimal separator in this locale.

Thus we get a null lcms profile from cmsCreateRGBProfile. Then as we pass this
null profile to lcms, which results in a segfault.

Many thanks to Alban Browaeys <prahal@yahoo.com> for debugging the problem.

Resolves: https://bugs.freedesktop.org/show_bug.cgi?id=58977