Ensure the color types are methods in GObject Introspection rather than functions
Many symbols are not named what GObject Introspection expects them to be so
we end up with a lot of functions in the global namespace instead of methods
e.g. "Cd.color_clear_xyz" instead of "Cd.ColorXYZ.clear".

There is no way to resolve this using annotations alone so the only way is to
create the symbols G-I expects. Ship a cd-deprecated.h header so we do not
break API or ABI. We'll remove this whenever we bump the soname next.

Based on a patch by Evan Nemerson <evan@coeus-group.com>, many thanks.