Do not return an error if a user manually adds a profile that has already been automatically added
This can happen if a calibration program creates a profile with an embedded
device-id which is then imported. The daemon automatically soft-adds the profile
to the device, and the calibration program or the user may then choose to hard
add the profile so that it is able to be set default.