Do not auto-add profiles due to device-id metadata if they have been manually removed by the user
If the user has manually removed the profile from the device, we should not keep
adding it at every fresh boot.

So we know this, set the timestamp to zero in the mapping database when the
profile is removed, and check for this value when soft-adding the profiles to a
device.

Also make a few more adjustments so that we ignore any mappings with a timestamp
of zero and add a self test for this functionality.