Add DTP94 native sensor support
The DTP94 is actually a really simple sensor. We don't know all the commands,
but it's trivial to capture the output using USB capture hardware when the
sensor is being used by ArgyllCMS. By replaying them we get back an XYZ reading.

We don't know what most of the commands actually do, but at least it works.