Fix timezone problem in colord_client_random_func test
colord interprets the “created time” metadata in a profile in the current
timezone, which is entirely reasonable. This means that the created time
in seconds-since-the-epoch-in-UTC will differ depending on the timezone
though, so we need to calculate the calendar time to compare in the test
rather than hardcoding it.

Prevents ‘make check’ from failing when the timezone is not equal to UTC