Ensure we can set the timestamp to zero when clearing a mapping database entry
This is the same bug we had with the device database. As we had no suitable
primary key, any 'INSERT OR REPLACE' operation would create additional entries
in the database, which would not be matched.

To mitigate this, create a suitable mapping table, and migrate all the users
data. Also include quite a few self tests, to ensure we do this correctly.