Save the common ICC profiles as GResources in the colord binary to reduce I/O
This reduces seek time a lot when colord is run with a cold cache. Each ICC
profile is typically only one block in size (< 4kb) and so we can reduce the
amount of seeking we do at startup for profiles we know will be installed.

This reduces the startup time by 35ms on my fast machine with an SSD, although
it'll make a *lot* more difference if you have a spinning disk hard drive.

If you're profiling this, remember to use 'echo 3 > /proc/sys/vm/drop_caches'
to ensure the hot file cache is flushed.