libcolord: Switch CdColor to using the GSlice allocator for a modest speedup
GSlice is designed for allocating a large number of small chunks of memory that
are the same size. By switching away from g_new0 (which is pretty much, but
slower than malloc) we can increase the speed of any program that allocates a
large number of color objects without breaking API or ABI.

Some broken users of the free functions that assumed the free-function was an
explicit GDestroyNotify have to be manually casted, but in the worst case this
is an easily detected compile warning that is easily fixed.

This commit also has the benefit that the _new() and _free() functions can be
used in languages that use the Colord-1.0 GIR, e.g. JavaScript and Python.