libcolord: Add functionality to interpolate arrays of CdColorRGB
This is required in a few different projects now, so abstract it out and do it
properly using Akima and falling back to cubic spline and then linear.

This also adds a hard dep on gsl, which is present on all distros, and is
already required by projects such as inkscape.

If anyone has a problem with this, let me know and I'll add a fallback to a
linear interpolation if the library is not available.