Use an XML source format for the generated profiles
Using cd-create-profile with ~12 arguments per call means we're not sure if the
data is in Yxy or xyY format, and also doesn't allow us to translate the
descriptions and copyright text information.

Also, using an XML format allows us to expand the amount of metadata we're using
in the future.