libcolord: Always write C-locale floating point values in IT8 files
LCMS uses 'sprintf(.10%g)' to convert floating point values to strings.
If LC_NUMERIC is set to something like 'nl_BE' then values such as 1.234 are
printed as '1,234'. This is against the CGATS standard and Argyll correctly
refuses to parse the file.

Wrap each call to get and set the floating point values with the safe GLib
functions of g_ascii_strtod() and g_ascii_dtostr().