Add a shim libcolordcompat library with some helpers for ArgyllCMS
ArgyllCMS uses WCS for Windows and ColorSync for OS-X to store the device to
profile mapping data. For situations where these are not available it also
provides the the UCMM functionality that stores the mapping as JSON data.

Providing a simple shim library allows ArgyllCMS to simply dlopen() the
/usr/lib64/libcolordcompat.so file if it exists and use colord for the mapping
data on Linux platforms.