main: Only syslog() profile additions when they're added via DBus
Currently we syslog 24 totally useless log messages about "Profile
added: <checksum>" on every boot.  This is a major offender in bootup
log spew.

I wrote this patch so that we do still log profile additions when a
client requests them via DBus, but we really shouldn't log about stuff
we just found on disk in the system ICC store.

Signed-off-by: Richard Hughes <richard@hughsie.com>