icc store: Don't create legacy locations
When cd_icc_store_search_kind is called with the
CD_ICC_STORE_SEARCH_FLAGS_CREATE_LOCATION flag, it creates
all the locations for the given kind. That is not really
ideal, since all but the first one are basically legacy,
and should not be forced into existence. The patch changes
things so that the create-location flag only affects the
first location.

https://bugs.freedesktop.org/show_bug.cgi?id=69765

Signed-off-by: Richard Hughes <richard@hughsie.com>