Use ICCv2 for the RGB working space profiles
We've discussed this at LGM, and there are still a significant number of
programs that do not work well with a v4 workflow. Much as I dislike using a
deprecated version this seems like a sensible compromise given we can use the
'dscm' tag for at least translated description on OSX and Linux.

We can revisit this when all applications do sane things with v4 profiles.
Test profiles remain v4, which I think is a good compromise.