Allow users to rename session and system ICC profiles
This does not actually modify the ICC for several reasons:

* This may be a system profile that other users may have set up mappings from

* Changing the ID or the MD5 of well-known profiles may break other applications

* Different users may want more or less technical descriptions

* We only get a readonly mapped file when creating profiles from the session

...so instead use the existing storage.db sqlite database to define per-uid
overrides for the 'Title' property. Users can 'modify' the title shown in
applications using colord by calling something like this:

$ colormgr profile-set-property icc-c2d76 Title 'Sony TV'

For applications using lcms2 directly to parse the file, the original
description will be shown. Additionally, there is no authentication step for
setting an overriden title on the profile, as it is only shown for that one
specific user. Titles can only be between 3 and 4096 chars in length and also
have to be valid UTF-8 strings.