Increase timeout to 60s for argyll spotread sampling
There was a 10s timeout for spotread to finish sampling.  This works
fine for light colors, in which case it may finish in as little as 3
seconds.  But the darker the color is, the longer it takes to finish.
When measuring black, it takes around 25-30 seconds for my Spyder 3
Express.

The timeout caused Gnome Control Panel to fail when attempting to
calibrate.