Check polkit version before defining g_autoptr macros
We cannot check on the internal name created by G_DECLARE_AUTO* macros
because it's not a pre-processor symbol.

Instead, we should check whether we're being built against a version of
PolKit that has those symbols, and only define them ourselves for
previous versions.

This commit fixes a build failure in GNOME Continuous.