data: fix owner & permissions on existing state files
If colord was previously configured to run as root, and subsequently
made to run as a different user, it will be unable to write to its state
files:

    colord[668]: CdDevice: failed to save mapping to database: SQL error: attempt to write a readonly database

'd' creates a directory if missing, and fixes its permissions if it
already exists, but it is not recursive. 'Z' recursively adjusts the
permission of the given directory and its children, but only if it
already exists. Together, these rules ensure that
@localstatedir@/lib/colord and @localstatedir@/lib/colord/icc both
exist, and that @localstatedir@/lib/colord and all its descendents are
owned by the correct user.