libcolord: Add cd_icc_create_default_full
The existing cd_icc_create_default creates a CdIcc with none of the load flags
set. A future change to gnome-settings-daemon is going to want to load an sRGB
profile with CD_ICC_LOAD_FLAGS_PRIMARIES in order to use it to construct a color
transformation matrix to plug into the window system.