libcolord: Add cd_mat33_is_finite
Move the logic to check whether any entries in a matrix are NaN or infinite from
cd_it8_utils_calculate_ccmx into a helper function in cd-math.c.

In addition, fix a bug in this code: if any entries of the matrix are exactly
zero, then fpclassify will return FP_ZERO instead of FP_NORMAL. Don't treat this
as a failure: only FP_NAN and FP_INFINITE are errors.