lib: avoid destructing LCMS plugin twice with lcms 2.14
lcms 2.14 contains a change to avoid a memory leak, but that change
assumes correct API usage. It's not necessary to both cmsUnregisterPluginsTHR()
and then cmsDeleteContext() -- we can just straight up delete the LCMS
context instead (cmsDeleteContext()).

So, follow upstream's suggestion & do that. This fixes memory corruption
when building colord on x86, for example.

Bug: https://github.com/mm2/Little-CMS/issues/344
Fixes: https://github.com/hughsie/colord/issues/145
Signed-off-by: Sam James <sam@gentoo.org>