lib/icc: Always close profile when loading fails
cd_icc_load() always takes ownership of the passed profile, but will
leak it if the initial condition fails and the funtion returns an error.
Only if the actual profile loading fails, will the passed profile
eventually not be leaked. Fix this by closing it immediately in the
first case.