Allow building with elogind
This makes the runtime systemd dep requirement optional when building
with -Dsystemd=true so that colord can be built with support for
libsystemd on distros that use elogind. There should be no impact here
when building colord on distros that use systemd, both libsystemd and
systemd should exist in that case.